requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10

# AI & ML
google-generativeai==0.3.2

# Data Sources
arxiv==2.1.0
scholarly==1.7.11
lxml==5.3.0

# Scheduling
APScheduler==3.10.4

# Database
sqlalchemy>=2.0.23

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Discord Integration
discord.py>=2.3.2

# FastAPI & Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Development
pytest==7.4.3
black==23.12.1
